+++ starter-code/views/boss/bossProfile.hbs
@@ -3,13 +3,13 @@
   
 <span>{{this.username}}:</span> 
 <p>{{this.role}}</p>
-<form action="/passport/{{this._id}}/deleteUser" method="POST">
+<form action="/boss/{{this._id}}/deleteUser" method="POST">
     <button type="submit">Delete</button>
 </form>
-<form action="/passport/{{this._id}}/edit" method="GET">
+<form action="/boss/{{this._id}}/edit" method="GET">
     <button type="submit">Edit</button>
 </form>
 {{/each}}
-<form action="/passport/createUser" method="GET">
+<form action="/boss/createUser" method="GET">
     <button type="submit">Create User</button>
 </form>
