+++ starter-code/views/boss/create.hbs
@@ -1,5 +1,5 @@
 <h1>Add new User</h1>
-<form action="/passport/newUser" method="POST">
+<form action="/boss/newUser" method="POST">
     <label>User name:</label>
     <input type="text" name="username">
     <label>Role:</label>
