--- starter-code/views/passport/profile.hbs
@@ -1,9 +1,0 @@
-<h1>Your profile</h1>
-<h3>{{this.username}}</h3>
-<p>{{this.role}}</p>
-<form action="/passport/{{this._id}}/edit" method="GET">
-    <button class="button" type="submit"> Edit Profile </button>
-</form>
-<form action="/passport/{{this._id}}/showProfiles" method="GET">
-    <button class="button" type="submit"> showProfiles </button>
-</form>
