+++ starter-code/views/user/createCourse.hbs
@@ -1,5 +1,5 @@
 <h1>Welcome!! Create a New Course</h1>
-<form action="/passport/createCourse" method="POST">
+<form action="/teacherA/create-course" method="POST">
     <label>Course Name: </label><input type="text" name="namecourse">
     <label>Professor: </label><input type="text" name="professor">
     <label>Duration</label><input type="text" name="duration">
