+++ starter-code/views/user/profile.hbs
@@ -1,0 +1,12 @@
+<h1>Your profile</h1>
+<h3>{{this.username}}</h3>
+<p>{{this.role}}</p>
+<form action="/user/{{this._id}}/edit" method="GET">
+    <button class="button" type="submit"> Edit Profile </button>
+</form>
+<form action="/user/{{this._id}}/show-profiles" method="GET">
+    <button class="button" type="submit"> showProfiles </button>
+</form>
+<form action="/user/view-courses" method="GET">
+     <button class="button" type="submit"> View Courses </button>
+</form>
