+++ starter-code/views/user/showCourses.hbs
@@ -1,6 +1,6 @@
 <h1>ALL Course Ironhack</h1>
 {{#each courses}}
    <label>Name:</label> <p>{{this.namecourse}}</p>
-   <label>Professor></label> <p>{{this.professor}}</p>
+   <label>Professor</label> <p>{{this.professor}}</p>
    <label>Duration</label> <p>{{this.duration}}</p>
 {{/each}}
