+++ starter-code/views/user/showProfiles.hbs
@@ -5,4 +5,4 @@
 
 {{/each}}
 
-<a href="/passportIndex">Back</a>
+<a href="/profiles">Back</a>
