+++ starter-code/views/user/teacherAProfile.hbs
@@ -1,15 +1,15 @@
 <h1>Your profile TA</h1>
 <h3>{{this.username}}</h3>
 <p>{{this.role}}</p>
-<form action="/passport/{{this._id}}/edit" method="GET">
+<form action="/user/{{this._id}}/edit" method="GET">
     <button class="button" type="submit"> Edit Profile </button>
 </form>
-<form action="/passport/{{this._id}}/showProfiles" method="GET">
+<form action="/user/{{this._id}}/show-profiles" method="GET">
     <button class="button" type="submit"> showProfiles </button>
 </form>
-<form action="/passport/createCourse" method="GET">
+<form action="/teacherA/create-course" method="GET">
      <button class="button" type="submit"> Create Course </button>
 </form>
-<form action="/passport/vieAllCourse" method="GET">
+<form action="/user/view-courses" method="GET">
      <button class="button" type="submit"> View Courses </button>
 </form>
